infrastructure/requirements.txt
attrs==25.3.0
aws-cdk-lib==2.202.0
aws-cdk.asset-awscli-v1==2.2.240
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-kinesisfirehose-alpha==2.186.0a0
aws-cdk.aws-kinesisfirehose-destinations-alpha==2.186.0a0
aws-cdk.cloud-assembly-schema==44.8.0
black==25.1.0
build==1.2.2.post1
cattrs==24.1.3
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
constructs==10.4.2
distlib==0.3.9
docformatter==1.7.7
filelock==3.18.0
flake8==7.3.0
identify==2.6.12
importlib_resources==6.5.2
isort==6.0.1
jsii==1.112.0
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pip-tools==7.4.1
platformdirs==4.3.8
pre_commit==4.2.0
publication==0.0.3
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
typeguard==2.13.3
typing_extensions==4.13.0
untokenize==0.1.1
virtualenv==20.31.2
wheel==0.45.1
